+++ packages/core/src/indexing/index.ts
@@ -219,9 +219,10 @@ export const createIndexing = ({
     }
   };
 
-  const executeEvent = async ({ event }: { event: Event }): Promise<void> => {
+  // metric label for "ponder_indexing_function_duration"
+  const executeEvent = async (event: Event): Promise<void> => {
     const indexingFunction = indexingFunctions[event.name];
-    const metricLabel = { event: event.name };
+    const metricLabel: { event: string } = { event: event.name };
 
     try {
       context.chain.id = event.chainId;
@@ -342,7 +343,7 @@ export const createIndexing = ({
           msg: `Started indexing function (event="${event.name}", checkpoint=${event.checkpoint})`,
         });
 
-        await executeEvent({ event });
+        await executeEvent(event);
 
         common.logger.trace({
           service: "indexing",
@@ -350,7 +351,6 @@ export const createIndexing = ({
         });
       }
 
-      // set completed events
       updateCompletedEvents();
     },
   };